rename classeValidator
