+++ api/src/cle/classe/classeValidator.ts
